requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
sqlparse
sqlvalidator
python-dotenv
pytest